kill squeezelite when closing the app
This change modifies the code to keep track of the squeezelite process
and to kill it when the app is closed.

The SqueezeLite struct makes it possible to expand this functionality
in the future to do other things such as restarting the squeezelite
process or implement multiple squeezelite instances.

This change has been tested on Linux.